"The dilemma of happiness" Poster
One of the most important themes in Mad Men is happiness, delivered by the concept of the American Dream. This poster focuses on the problem that people are facing when seeking happiness, using examples from the shows and real advertisements from the 1960s. This dilemma is precisely a never-ending loop consists of expectation, disappointment, emptiness.